requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas>=2.2.0
lxml>=4.10.0
supabase>=2.2.0
numpy
python-dotenv
google-cloud-bigquery
pandas-gbq
db-dtypes